apps/parser/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.0
python-docx==1.1.0
lxml==4.9.3
Pillow==10.2.0

beautifulsoup4==4.12.3
olefile==0.46
latex2mathml==3.78.0
mathml2omml==0.0.2
